Refactor: split manifest into own file, verify race-safety
Maintainability improvements:
- Extract Manifest() (240+ lines of JSON schemas and templates) into
  manifest.go, keeping expedia.go focused on struct, do(), and auth
- Fix trailing blank line in const block
- Update README file structure to reflect new layout
- Verified all tests pass with -race flag (no race conditions)
